drafts: Move some functions from draft_overlay_ui.
Preparation for an upcoming commit.
